requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==0.8.11
